fix(mobile): fixed convertToValue

+++ mobilesdk/sdk/sdk.go
@@ -226,8 +226,8 @@ func (s *StorageSDK) GetReadPoolInfo(clientID string) (string, error) {
 func (s *StorageSDK) WritePoolLock(durInSeconds int64, tokens, fee float64, allocID string) error {
 	_, _, err := sdk.WritePoolLock(
 		allocID,
-		uint64(zcn.ConvertToValue(tokens)),
-		uint64(zcn.ConvertToValue(fee)))
+		zcn.ConvertToValue(tokens),
+		zcn.ConvertToValue(fee))
 	return err
 }
 
@@ -312,5 +312,5 @@ func decodeTicket(ticket string) (string, string, uint64, error) {
 	markerStr, _ := json.Marshal(markerInput)
 
 	s, _ := strconv.ParseFloat(string(fmt.Sprintf("%v", lock)), 64)
-	return string(recipientPublicKey), string(markerStr), uint64(zcn.ConvertToValue(s)), nil
+	return string(recipientPublicKey), string(markerStr), zcn.ConvertToValue(s), nil
 }
